+++ nextstep-mvc/src/main/java/nextstep/mvc/exception/AnnotationScanFailException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.exception;
 
 public class AnnotationScanFailException extends RuntimeException {
     private static final String ANNOTAION_SCAN_FAIL_MESSAGE = "해당 어노테이션을 찾지 못했습니다.";
