+++ nextstep-mvc/src/main/java/nextstep/mvc/exception/HandlerAdapterNotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.exception;
 
 public class HandlerAdapterNotFoundException extends RuntimeException {
     private static final String HANDLER_ADAPTER_NOT_FOUND_MESSAGE = "어댑터를 찾을 수 없습니다.";
