+++ nextstep-mvc/src/main/java/nextstep/mvc/exception/HandlerNotExistException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.exception;
 
 public class HandlerNotExistException extends RuntimeException {
     private static final String HANDLER_NOT_EXIST_MESSAGE = "핸들러가 존재하지 않습니다.";
