+++ nextstep-mvc/src/main/java/nextstep/mvc/exception/HandlerNotFoundException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.exception;
 
 public class HandlerNotFoundException extends RuntimeException {
     private static final String HANDLER_NOT_FOUND_MESSAGE = "경로에 해당하는 핸들러가 존재하지 않습니다.";
