+++ nextstep-mvc/src/main/java/nextstep/mvc/exception/ObjectMapperException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.exception;
 
 public class ObjectMapperException extends RuntimeException {
     public ObjectMapperException() {
