+++ nextstep-mvc/src/main/java/nextstep/mvc/exception/UnsupportedViewException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.exception;
 
 public class UnsupportedViewException extends RuntimeException {
     private static final String UNSUPPORTED_VIEW_MESSAGE = "지원하지 않는 뷰입니다.";
