+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/HandlerMethodArgumentResolverTest.java
@@ -42,8 +42,8 @@ public class HandlerMethodArgumentResolverTest {
 
     private Method getMethod(String name, Method[] methods) {
         return Arrays.stream(methods)
-                .filter(method -> method.getName().equals(name))
-                .findFirst()
-                .get();
+            .filter(method -> method.getName().equals(name))
+            .findFirst()
+            .get();
     }
 }
