+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/MyController.java
@@ -26,10 +26,10 @@ public class MyController {
     @RequestMapping(value = "/users", method = RequestMethod.POST)
     public ModelAndView save(HttpServletRequest request, HttpServletResponse response) {
         User user = new User(
-                request.getParameter("userId"),
-                request.getParameter("password"),
-                request.getParameter("name"),
-                request.getParameter("email"));
+            request.getParameter("userId"),
+            request.getParameter("password"),
+            request.getParameter("name"),
+            request.getParameter("email"));
         logger.debug("User : {}", user);
         DataBase.addUser(user);
         return null;
