+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/TestUser.java
@@ -26,9 +26,9 @@ public class TestUser {
     @Override
     public String toString() {
         return "TestUser{" +
-                "userId='" + userId + '\'' +
-                ", password='" + password + '\'' +
-                ", age=" + age +
-                '}';
+            "userId='" + userId + '\'' +
+            ", password='" + password + '\'' +
+            ", age=" + age +
+            '}';
     }
 }
