+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/User.java
@@ -77,11 +77,8 @@ public class User {
         } else if (!name.equals(other.name))
             return false;
         if (userId == null) {
-            if (other.userId != null)
-                return false;
-        } else if (!userId.equals(other.userId))
-            return false;
-        return true;
+            return other.userId == null;
+        } else return userId.equals(other.userId);
     }
 
     @Override
